+++ components/org.wso2.carbon.identity.api.server.action.management/org.wso2.carbon.identity.api.server.action.management.v1/pom.xml
@@ -34,7 +34,7 @@
     <dependencies>
         <dependency>
             <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.action.mgt</artifactId>
+            <artifactId>org.wso2.carbon.identity.action.management</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
